Shuffle around files to create RGSS library properly
Can't easily dynamically link the RGSS library with Ruby in the game executable. So instead, two separate dynamic libraries are created - one for the native extension (binding) and one for the game engine. Static Ruby and the RPG namespace has been moved into the engine. The game executable still needs to be cleaned up, as it's referencing internals of the engine that it shouldn't.